+++ assets/lambda/src/event-store/create-system-event-store-event-service.ts
@@ -3,24 +3,7 @@ import {
     SystemEventBridgeEvent,
     SystemEventStorePushEvent,
 } from '../events/user-event';
-
-export interface SystemEventEntity {
-    id: string;
-    version: number;
-    event: SystemEventStorePushEvent;
-    eventType: string;
-    sourceSystem: string;
-    time: string;
-}
-
-export const systemEventEntitySchema = {
-    type: 'object',
-    properties: {
-        id: { type: 'string' },
-    },
-    required: ['id'],
-    additionalProperties: true,
-};
+import { SystemEventEntity } from './system-event-entity';
 
 export class StoreSystemEventService {
     constructor(private repository: SystemEventStoreRepository) {}
