+++ assets/lambda/src/event-store/system-event-entity.ts
@@ -1,0 +1,19 @@
+import { SystemEventStorePushEvent } from '../events/user-event';
+
+export interface SystemEventEntity {
+    id: string;
+    version: number;
+    event: SystemEventStorePushEvent;
+    eventType: string;
+    sourceSystem: string;
+    time: string;
+}
+
+export const systemEventEntitySchema = {
+    type: 'object',
+    properties: {
+        id: { type: 'string' },
+    },
+    required: ['id'],
+    additionalProperties: true,
+};
